+++ tablero.py
@@ -1,38 +1,8 @@
 
 import pygame, sys
 from constantes import *
+from BotonBloque import BotonBloque
         
-
-
-class blockButton():
-    def __init__(self, rect, fila, columna, blockMatrix):
-        self.rect = rect
-        self.clicked = False
-        self.fila = fila
-        self.columna = columna
-        self.blockMatrix = blockMatrix
-       # self.value = 1 # 1 marcado, 2 cruz
-
-    def draw(self, grilla):
-        #self.marcar(value)
-        # value: 0 vacio, 1 marcado, 2 cruz
-
-       # print(np.matrix(self.matrizValoresBloques))
-        #print("fil: ", self.fila, " col: ", self.columna)
-        valorEnMatriz = self.matrizValoresBloques[self.fila][self.columna]
-
-        if valorEnMatriz == 1:         # color para indicar que esta marcado
-            pygame.draw.rect(grilla, DARK_BLUE, self.rect) 
-        
-        elif valorEnMatriz == 2:       # color para indicar que esta tachado
-            pygame.draw.rect(grilla, DARK_BEIGE, self.rect) 
-        
-        elif valorEnMatriz == 0:
-            pygame.draw.rect(grilla, BEIGE, self.rect)
-
-        pygame.display.update()
-       
-
 
 def drawGrid(screen, blockCant, matriz):
     
@@ -139,7 +109,7 @@ def manejarEventos(self, blockSize, matrizValoresBloques):
 def etapaTablero(screen, blockCant, matriz):
     pygame.display.set_caption('Tablero')
    
-    drawNumberIndicators(screen, blockCant, matriz)
+    
     drawGrid(screen, blockCant, matriz)
